+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/deepseek
 
+## 0.1.3
+
+### Patch Changes
+
+- 361fd08: chore: update a few add'l processor references to extractor
+- Updated dependencies [361fd08]
+  - @ai-sdk/openai-compatible@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes